requirements.txt
fastapi>=0.110.0
pydantic>=2.6.3
uvicorn>=0.27.1
sqlalchemy>=2.0.27
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.9
alembic>=1.13.1
psycopg2-binary>=2.9.1,<2.10.0
sendgrid>=6.9.0,<7.0.0
jinja2>=3.0.1,<3.1.0
python-dotenv>=1.0.1
httpx>=0.27.0
beautifulsoup4>=4.9.0,<4.10.0
fastapi-mail>=1.4.1
email-validator>=1.1.3
pytest>=8.0.2
bcrypt>=4.1.2
gunicorn>=21.2.0
sentry-sdk[fastapi]>=1.40.0
slowapi>=0.1.9
aiohttp>=3.9.3
supabase>=2.3.4
backoff>=2.2.1 